cmd/k8s-operator: split operator.go into svc.go/sts.go
Updates #502
